// src/types/css-props.ts
export interface SizeProps {
  height: string;
  width: string;
}

export interface ColorProps {
  greenPrimary: string;
  greenSecondary: string;
  greenFocus: string;
  greenLight: string;
  greenLightBg: string;
  whitePrimaryBg: string;
  greyLight: string;
  greyNeutral: string;
  orangePrimary: string;
  orangeLight: string;
  semiBlack: string;
}